+++ crates/schematic/src/schema/mod.rs
@@ -22,7 +22,7 @@ pub use renderers::jsonc_template::*;
 /// Helpers for config templates.
 #[cfg(feature = "renderer_template")]
 #[allow(deprecated)]
-pub use renderers::template::{self, TemplateOptions, TemplateRenderer};
+pub use renderers::template::{self, TemplateOptions};
 
 /// Renders TOML config templates.
 #[cfg(all(feature = "renderer_template", feature = "toml"))]